Estado de Ordenes

+++ application/models/Model_orders.php
@@ -1,12 +1,10 @@
 <?php 
-
 class Model_orders extends CI_Model
 {
 	public function __construct()
 	{
 		parent::__construct();
 	}
-
 	/* get the orders data */
 	public function getOrdersData($id = null)
 	{
@@ -15,30 +13,37 @@ class Model_orders extends CI_Model
 			$query = $this->db->query($sql, array($id));
 			return $query->row_array();
 		}
-
 		$sql = "SELECT * FROM orders ORDER BY id DESC";
 		$query = $this->db->query($sql);
 		return $query->result_array();
 	}
 
+	public function getOrdersData2($id = null)
+	{
+		if($id) {
+			$sql = "SELECT * FROM orders WHERE  id = ? and(  estado_orden='0' or   estado_orden='1') ";
+			$query = $this->db->query($sql, array($id));
+			return $query->row_array();
+		}
+
+		$sql = "SELECT * FROM orders WHERE estado_orden='0' or   estado_orden='1'  ORDER BY id DESC";
+		$query = $this->db->query($sql);
+		return $query->result_array();
+	}
 	// get the orders item data
 	public function getOrdersItemData($order_id = null)
 	{
 		if(!$order_id) {
 			return false;
 		}
-
 		$sql = "SELECT * FROM orders_item WHERE order_id = ?";
 		$query = $this->db->query($sql, array($order_id));
 		return $query->result_array();
 	}
-
 	public function create()
 	{
 		$name_mesa = $this->input->post('id_mesa');
-
 		$this->load->model('model_mesas');
-
 		$mesa = $this->model_mesas->getMesaforName($name_mesa) ;
 			  
 		// ****** MODIFICAR ESTADO DE MESA
@@ -51,14 +56,7 @@ class Model_orders extends CI_Model
 				else {  // $mesa['active']==2  // OCUPADO
 //// No se puede crear orden 
 // MANDAR A EDIT MESA 
-
-
-
 				}
-
-
-
-
 		$user_id = $this->session->userdata('id');
 		$bill_no = 'FISI-'.strtoupper(substr(md5(uniqid(mt_rand(), true)), 0, 4));
     	$data = array(
@@ -80,15 +78,11 @@ class Model_orders extends CI_Model
     		'estado_orden' => 0
 			
     	);
-
  
-
 		// ******
 		$insert = $this->db->insert('orders', $data);
 		$order_id = $this->db->insert_id();
-
 		$this->load->model('model_products');
-
 		$count_product = count($this->input->post('product'));
     	for($x = 0; $x < $count_product; $x++) {
     		$items = array(
@@ -98,23 +92,15 @@ class Model_orders extends CI_Model
     			'rate' => $this->input->post('rate_value')[$x],
     			'amount' => $this->input->post('amount_value')[$x],
     		);
-
     		$this->db->insert('orders_item', $items);
-
     		// now decrease the stock from the product
     		$product_data = $this->model_products->getProductData($this->input->post('product')[$x]);
     		$qty = (int) $product_data['qty'] - (int) $this->input->post('qty')[$x];
-
     		$update_product = array('qty' => $qty);
-
-
     		$this->model_products->update($update_product, $this->input->post('product')[$x]);
     	}
-
 		return ($order_id) ? $order_id : false;
 	}
-
-
 	public function countOrderItem($order_id)
 	{
 		if($order_id) {
@@ -123,13 +109,11 @@ class Model_orders extends CI_Model
 			return $query->num_rows();
 		}
 	}
-
 	public function update($id)
 	{
 		if($id) {
 			$user_id = $this->session->userdata('id');
 			// fetch the order data 
-
 			$data = array(
 				'customer_name' => $this->input->post('customer_name'),
 	    		'customer_address' => $this->input->post('customer_address'),
@@ -144,10 +128,8 @@ class Model_orders extends CI_Model
 	    		'paid_status' => $this->input->post('paid_status'),
 	    		'user_id' => $user_id
 	    	);
-
 			$this->db->where('id', $id);
 			$update = $this->db->update('orders', $data);
-
 			// now the order item 
 			// first we will replace the product qty to original and subtract the qty again
 			$this->load->model('model_products');
@@ -163,11 +145,9 @@ class Model_orders extends CI_Model
 				// update the product qty
 				$this->model_products->update($update_product_data, $product_id);
 			}
-
 			// now remove the order item data 
 			$this->db->where('order_id', $id);
 			$this->db->delete('orders_item');
-
 			// now decrease the product qty
 			$count_product = count($this->input->post('product'));
 	    	for($x = 0; $x < $count_product; $x++) {
@@ -179,32 +159,102 @@ class Model_orders extends CI_Model
 	    			'amount' => $this->input->post('amount_value')[$x],
 	    		);
 	    		$this->db->insert('orders_item', $items);
-
 	    		// now decrease the stock from the product
 	    		$product_data = $this->model_products->getProductData($this->input->post('product')[$x]);
 	    		$qty = (int) $product_data['qty'] - (int) $this->input->post('qty')[$x];
-
 	    		$update_product = array('qty' => $qty);
 	    		$this->model_products->update($update_product, $this->input->post('product')[$x]);
 	    	}
-
 			return true;
 		}
 	}
 
 
 
+	public function updateEs($id)
+	{
+		if($id) {
+			$user_id = $this->session->userdata('id');
+			// fetch the order data 
+			$data = array(
+				'estado_orden'=>$this->input->post('estado_orden'),
+				// 'customer_name' => $this->input->post('customer_name'),
+	    		// 'customer_address' => $this->input->post('customer_address'),
+	    		// 'customer_phone' => $this->input->post('customer_phone'),
+	    		// 'gross_amount' => $this->input->post('gross_amount_value'),
+	    		// 'service_charge_rate' => $this->input->post('service_charge_rate'),
+	    		// 'service_charge' => ($this->input->post('service_charge_value') > 0) ? $this->input->post('service_charge_value'):0,
+	    		// 'vat_charge_rate' => $this->input->post('vat_charge_rate'),
+	    		// 'vat_charge' => ($this->input->post('vat_charge_value') > 0) ? $this->input->post('vat_charge_value') : 0,
+	    		// 'net_amount' => $this->input->post('net_amount_value'),
+	    		// 'discount' => $this->input->post('discount'),
+	    		// 'paid_status' => $this->input->post('paid_status'),
+	    		// 'user_id' => $user_id
+	    	);
+			$this->db->where('id', $id);
+			$update = $this->db->update('orders', $data);
+			// now the order item 
+			// first we will replace the product qty to original and subtract the qty again
+			$this->load->model('model_products');
+			$get_order_item = $this->getOrdersItemData($id);
+			foreach ($get_order_item as $k => $v) {
+				$product_id = $v['product_id'];
+				$qty = $v['qty'];
+				// get the product 
+				$product_data = $this->model_products->getProductData($product_id);
+				$update_qty = $qty + $product_data['qty'];
+				$update_product_data = array('qty' => $update_qty);
+				
+				// update the product qty
+				$this->model_products->update($update_product_data, $product_id);
+			}
+			// now remove the order item data 
+			$this->db->where('order_id', $id);
+			$this->db->delete('orders_item');
+			// now decrease the product qty
+			$count_product = count($this->input->post('product'));
+	    	for($x = 0; $x < $count_product; $x++) {
+	    		$items = array(
+	    			'order_id' => $id,
+	    			'product_id' => $this->input->post('product')[$x],
+	    			'qty' => $this->input->post('qty')[$x],
+	    			'rate' => $this->input->post('rate_value')[$x],
+	    			'amount' => $this->input->post('amount_value')[$x],
+	    		);
+	    		$this->db->insert('orders_item', $items);
+	    		// now decrease the stock from the product
+	    		$product_data = $this->model_products->getProductData($this->input->post('product')[$x]);
+	    		$qty = (int) $product_data['qty'] - (int) $this->input->post('qty')[$x];
+	    		$update_product = array('qty' => $qty);
+	    		$this->model_products->update($update_product, $this->input->post('product')[$x]);
+	    	}
+			return true;
+		}
+	}
+
+
 	public function remove($id)
 	{
 		if($id) {
-			$this->db->where('id', $id);
-			$delete = $this->db->delete('orders');
 
-			$this->db->where('order_id', $id);
-			$delete_item = $this->db->delete('orders_item');
-			return ($delete == true && $delete_item) ? true : false;
+			$order = $this->model_orders->getOrdersData($id) ;
+			
+            if($order['estado_orden']==0){
+				$this->db->where('id', $id);
+				$delete = $this->db->delete('orders');
+				$this->db->where('order_id', $id);
+				$delete_item = $this->db->delete('orders_item');
+				return ($delete == true && $delete_item) ? true : false;
+
+			}else{
+				return false;
+			}
+		
+
+		
 		}
 	}
+
 
 	public function countTotalPaidOrders()
 	{
@@ -212,5 +262,4 @@ class Model_orders extends CI_Model
 		$query = $this->db->query($sql, array(1));
 		return $query->num_rows();
 	}
-
 }
